\name{breastCancer}
\docType{data}
\alias{breastCancer}
\title{Breast Cancer Wisconsin (Original) Data Set}
\description{ Original Wisconsin Breast Cancer Database
}
\usage{breastCancer}
\format{A matrix containing 699 observations and 10 attributes with missing values.}
\source{ http://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+(Original)}
\references{Wolberg, W.H., and Mangasarian, O.L. (1990). Multisurface method of pattern separation for medical diagnosis applied to breast cytology. In Proceedings of the National Academy of Sciences, 87, 9193--9196. 

Zhang, J. (1992). Selecting typical instances in instance-based learning. In Proceedings of the Ninth International Machine Learning Conference (pp. 470--479). Aberdeen, Scotland: Morgan Kaufmann. 
}