\name{CarEvaluation}
\docType{data}
\alias{carEvaluation}
\title{Car Evaluation Data Set}
\description{
  Derived from simple hierarchical decision model, this database may be useful for testing constructive induction and structure discovery methods.
}
\usage{carEvaluation}
\format{A matrix containing 1728 observations and 6 attributes.}
\source{ http://archive.ics.uci.edu/ml/datasets/Car+Evaluation}
\references{ M. Bohanec and V. Rajkovic: Knowledge acquisition and explanation for multi-attribute decision making. In 8th Intl Workshop on Expert Systems and their Applications, Avignon, France. pages 59-78, 1988. 

B. Zupan, M. Bohanec, I. Bratko, J. Demsar: Machine learning by function decomposition. ICML-97, Nashville, TN. 1997 (to appear) 
}