\name{ConcreteCompressiveStrength}
\docType{data}
\alias{ConcreteCompressiveStrength}
\title{Concrete Compressive Strength Data Set}
\description{
  Concrete is the most important material in civil engineering. The concrete compressive strength is a highly nonlinear function of age and ingredients.
}
\usage{ConcreteCompressiveStrength}
\format{A matrix containing 1030 observations and 9 attributes.}
\source{ http://archive.ics.uci.edu/ml/datasets/Concrete+Compressive+Strength}
\references{ I-Cheng Yeh, "Modeling of strength of high performance concrete using artificial neural networks," Cement and Concrete Research, Vol. 28, No. 12, pp. 1797-1808 (1998). 
}