\name{internalFunctions}
\alias{CheckSameValuesInAllAttributes} \alias{CheckSameValuesInLabels} \alias{HuberDist} \alias{Id}	\alias{L1Dist}	\alias{L1DistCPP}	
\alias{L1InformationGainCPP} \alias{L1AsymetricInformationGainCPP} \alias{L2AsymetricInformationGainCPP} \alias{LInfCPP} \alias{asymetricCrossEntropyCPP} \alias{asymetricGiniCPP} \alias{asymetricInformationGainCPP} \alias{L2.logDist}	\alias{L2Dist}	\alias{L2DistCPP} \alias{L2InformationGainCPP}	\alias{NAFeatures}
\alias{NAfactor2matrix}	\alias{OOBquantiles} \alias{XMinMaxCPP}	\alias{as.true.matrix} \alias{bCICore}	\alias{category2Proba} \alias{category2Quantile}	\alias{categoryCombination}	\alias{checkUniqueObsCPP} \alias{classifyCPP}	
\alias{classifyMatrixCPP}	\alias{combineRUFObjects}	\alias{concat} \alias{concatCore}	\alias{conditionalCrossEntropyCPP}	\alias{conditionalGiniCPP} \alias{confusion.matrix}	\alias{count.factor}	\alias{crossEntropyCPP}
\alias{dates2numeric}  
\alias{which.is.duplicate} \alias{define_train_test_sets}	\alias{difflog}	\alias{dummy.recode} \alias{entropyInformationGainCPP}	\alias{estimatePredictionAccuracy} 	\alias{expectedSquaredBias}	\alias{extractYFromData} \alias{fScore}	\alias{factor2matrix}	\alias{factor2vector}	\alias{fillVariablesNames} \alias{fillWith}	\alias{filter.forest}	\alias{filter.object}	\alias{find.first.idx} \alias{find.idx}	\alias{find.root}	\alias{fullNDCG}	\alias{fullNode}	\alias{gMean} \alias{generalization.error}	\alias{generic.log}	\alias{generic.smoothing.log} \alias{genericCbind}	\alias{genericNode}	\alias{genericOutput}	\alias{getCorr}	\alias{getOddEven} \alias{getVotesProbability}	\alias{getVotesProbability2}	\alias{giniCPP} \alias{imputeCategoryForTestData}	\alias{inDummies}	
\alias{insert.in.vector} \alias{insert.in.vector2}	\alias{is.wholenumber}	\alias{keep.index}	\alias{lagFunction}
\alias{leafNode}	\alias{localTreeImportance}	\alias{localVariableImportance} \alias{majorityClass}	\alias{matrix2factor}	\alias{mergeLists}	\alias{min_or_max}	\alias{modX} \alias{monitorOOBError}	\alias{na.impute}	\alias{na.missing}	\alias{na.replace}	\alias{ndcg} \alias{onlineClassify}	\alias{onlineCombineRUF}	\alias{optimizeFalsePositives} \alias{options.filter}	\alias{outputPerturbationSampling}	\alias{outsideConfIntLevels} \alias{overSampling}	\alias{parallelNA.replace}	\alias{permuteCatValues}
\alias{perspWithcol}	
\alias{plotTreeCore}	\alias{plotTreeCore2}	\alias{predictDecisionTree} \alias{predictionvsResponses}	\alias{pseudoHuberDist}	\alias{pseudoNAReplace} \alias{rUniformForest.merge}	\alias{rUniformForestPredict}	\alias{randomCombination}
\alias{randomUniformForestCore}	\alias{randomUniformForestCore.big}	\alias{randomUniformForestCore.merge}
\alias{randomUniformForestCore.predict}	\alias{randomWhichMax}	\alias{randomize} \alias{rankingTrainData} \alias{residualsRandomUniformForest}	\alias{rm.InAList} \alias{rm.correlation}	\alias{rm.string}	\alias{rm.tempdir}	\alias{rmInAListByNames} \alias{rmInf}
\alias{rmNA} 
\alias{rmNoise} \alias{rollApplyFunction}	\alias{runifMatrixCPP} \alias{sampleDirichlet}	\alias{scale2AnyValues}	\alias{setManyDatasets}	\alias{smoothing.log} \alias{someErrorType}	\alias{sortCPP}	\alias{sortDataframe}	\alias{sortMatrix}	\alias{standardize}
\alias{standardize_vect} \alias{strength_and_correlation} \alias{timer}	\alias{twoColumnsImportance}
\alias{uniformDecisionTree}	\alias{vector2factor}	\alias{vector2matrix}	\alias{weightedVote} \alias{weightedVoteModel}	\alias{which.is.factor}	\alias{which.is.na} \alias{which.is.wholenumber} \alias{myAUC} \alias{matrix2factor2} \alias{NATreatment}
\alias{filterOutliers} 
\alias{which.is.nearestCenter} \alias{variance} \alias{interClassesVariance} \alias{intraClassesVariance}  \alias{kBiggestProximities} \alias{gap.stats} \alias{kMeans} \alias{hClust} \alias{proximitiesMatrix} \alias{observationsImportance} \alias{A2Rplot} \alias{A2Rplot.default} \alias{A2Rplot.hclust} \alias{cutree.order}  \alias{splitVarCore} \alias{timeStampCore}
\alias{MDSscale} \alias{mergeOutliers} \alias{unsupervised2supervised} \alias{scalingMDS} \alias{rewind.trees} \alias{reduce.trees} \alias{updateCombined.unsupervised}
\alias{specClust} \alias{rm.coordinates} \alias{OOBVotesScale} \alias{kappaStat} \alias{copulaLike} 
\alias{subsampleFile} \alias{modelingResiduals} \alias{estimaterequiredSampleSize} \alias{estimatePredictionAccuracy} 
	
\title{All internal functions}

\description{
	Internal functions for random Uniform Forests
}