\name{wineQualityRed}
\docType{data}
\alias{wineQualityRed}
\title{Wine Quality Data Set}
\description{
  Two datasets are included, related to red and white vinho verde wine samples, from the north of Portugal. The goal is to model wine quality based on physicochemical tests.
}
\usage{wineQualityRed}
\format{A matrix (for red wine) containing 1599 observations and 12 attributes.}
\source{ http://archive.ics.uci.edu/ml/datasets/Wine+Quality}
\references{ Paulo Cortez, University of Minho, Guimaraes, Portugal, http://www3.dsi.uminho.pt/pcortez 
A. Cerdeira, F. Almeida, T. Matos and J. Reis, Viticulture Commission of the Vinho Verde Region(CVRVV), Porto, Portugal 
2009

P. Cortez, A. Cerdeira, F. Almeida, T. Matos and J. Reis. Modeling wine preferences by data mining from physicochemical properties.In Decision Support Systems, Elsevier, 47(4):547-553, 2009.
}